feat: functions select in manual run
